llamaindex-docs-rag/requirements.txt
requests
beautifulsoup4
llama-index
openai
pinecone-client
python-dotenv
setuptools
black
watchdog
fastapi
uvicorn
llmlingua
accelerate